chore: consolidate test structure
