--- pages/api/products.js.orig
@@ -1,7 +1,0 @@
-// pages/api/products.js
-import { getProducts } from '../../lib/products';
-
-export default async function handler(req, res) {
-  const products = await getProducts();
-  res.status(200).json(products);
-}
