--- pages/index.js.rej
@@ -1,10 +1,0 @@
---- pages/index.js
-+++ pages/index.js
-@@ -1,6 +1,7 @@
- import React, { useState } from 'react';
- import products from '../data/products.json';
- 
- const HomePage = () => {
-+  const [searchTerm, setSearchTerm] = useState('');
-   
-   return (
