+++ tests/attributeSuggestor.test.js
