+++ tests/attributes-core.test.js
