+++ tests/inStockFilter.test.js
