+++ tests/lib/attributeGroups.test.js
