+++ tests/manual/attribute-groups-api.test.js
