+++ tests/slugify.test.js
